[FE-182]refactor: 파일사이즈 함수 외부로 -재사용을 위해 파일 사이즈 구하는 함수 utils로 분리
